+++ controllers/serviceexport/reconciler.go
@@ -31,8 +31,8 @@ import (
 
 	kubeslicev1beta1 "github.com/kubeslice/worker-operator/api/v1beta1"
 	"github.com/kubeslice/worker-operator/controllers"
-	"github.com/kubeslice/worker-operator/internal/logger"
 	"github.com/kubeslice/worker-operator/pkg/events"
+	"github.com/kubeslice/worker-operator/pkg/logger"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 )
