+++ controllers/slice/reconciler.go
@@ -34,9 +34,9 @@ import (
 
 	kubeslicev1beta1 "github.com/kubeslice/worker-operator/api/v1beta1"
 	"github.com/kubeslice/worker-operator/controllers"
-	"github.com/kubeslice/worker-operator/internal/logger"
-	"github.com/kubeslice/worker-operator/internal/manifest"
 	"github.com/kubeslice/worker-operator/pkg/events"
+	"github.com/kubeslice/worker-operator/pkg/logger"
+	"github.com/kubeslice/worker-operator/pkg/manifest"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 )
