+++ pkg/cluster/node.go
@@ -21,7 +21,7 @@ package cluster
 import (
 	"context"
 	"fmt"
-	"github.com/kubeslice/worker-operator/internal/logger"
+	"github.com/kubeslice/worker-operator/pkg/logger"
 	corev1 "k8s.io/api/core/v1"
 	"os"
 	"sigs.k8s.io/controller-runtime/pkg/client"
