+++ pkg/hub/hubclient/hubclient.go
@@ -36,8 +36,8 @@ import (
 	hubv1alpha1 "github.com/kubeslice/apis/pkg/controller/v1alpha1"
 	spokev1alpha1 "github.com/kubeslice/apis/pkg/worker/v1alpha1"
 	kubeslicev1beta1 "github.com/kubeslice/worker-operator/api/v1beta1"
-	"github.com/kubeslice/worker-operator/internal/cluster"
-	"github.com/kubeslice/worker-operator/internal/logger"
+	"github.com/kubeslice/worker-operator/pkg/cluster"
+	"github.com/kubeslice/worker-operator/pkg/logger"
 )
 
 var scheme = runtime.NewScheme()
