+++ pkg/manifest/file.go
@@ -27,7 +27,7 @@ import (
 
 	"encoding/json"
 
-	"github.com/kubeslice/worker-operator/internal/logger"
+	"github.com/kubeslice/worker-operator/pkg/logger"
 )
 
 var log = logger.NewLogger()
