+++ tests/hub/hub_suite_test.go
@@ -38,8 +38,8 @@ import (
 
 	spokev1alpha1 "github.com/kubeslice/apis/pkg/worker/v1alpha1"
 	kubeslicev1beta1 "github.com/kubeslice/worker-operator/api/v1beta1"
-	"github.com/kubeslice/worker-operator/internal/hub/controllers"
 	"github.com/kubeslice/worker-operator/pkg/events"
+	"github.com/kubeslice/worker-operator/pkg/hub/controllers"
 	"sigs.k8s.io/controller-runtime/pkg/predicate"
 )
