+++ tests/spoke/spoke_suite_test.go
@@ -42,10 +42,10 @@ import (
 	"github.com/kubeslice/worker-operator/controllers/serviceimport"
 	"github.com/kubeslice/worker-operator/controllers/slice"
 	"github.com/kubeslice/worker-operator/controllers/slicegateway"
-	hub "github.com/kubeslice/worker-operator/internal/hub/hubclient"
-	namespace "github.com/kubeslice/worker-operator/internal/namespace/controllers"
-	"github.com/kubeslice/worker-operator/internal/networkpolicy"
 	"github.com/kubeslice/worker-operator/pkg/events"
+	hub "github.com/kubeslice/worker-operator/pkg/hub/hubclient"
+	namespace "github.com/kubeslice/worker-operator/pkg/namespace/controllers"
+	"github.com/kubeslice/worker-operator/pkg/networkpolicy"
 	hce "github.com/kubeslice/worker-operator/tests/emulator/hubclient"
 	workernetop "github.com/kubeslice/worker-operator/tests/emulator/workerclient/netop"
 	workerrouter "github.com/kubeslice/worker-operator/tests/emulator/workerclient/router"
